% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logwatch.R
\name{logwatch}
\alias{logwatch}
\title{Watch logs}
\usage{
logwatch(what, get_status, get_log, show_log = TRUE, poll = 1)
}
\arguments{
\item{what}{Very short summary of the process underlying the
logs. Examples include "task", "installation", etc.}

\item{get_status}{Function that will fetch a status. Accepts no
arguments, and returns one of a set of "waiting", "running" or
any status you like after that.  When status is "waiting" we
show a spinner but do not stream logs. When "running" we stream
logs (but don't show a spinner). Any other statement returns.}

\item{get_log}{A callback to read logs of the installation
(something like \code{function() readLines(filename, warn = FALSE)}
may be sufficient)}

\item{show_log}{Logical, indicating if the installation log should
be printed}

\item{poll}{Time, in seconds, used to throttle calls to the status
function. The default is 1 second}
}
\value{
A list with elements:
\itemize{
\item status: Your final status call
\item start: The start time
\item end: The end time
}
}
\description{
Watch for a log of something.
}
